Refinement to SoD druid means testing
SoD druids lack the higher-level perks, so the detection for these in SoD would fail. I initially ignored this as I didn't think
druids reached levels in SoD to get the new perks, but I double-checked and was just wrong. Cheers to yuyuanshuo for pointing
this out in the commit comments.