Better Enhanced Bard Song treatment, better RR compat
A clever bit of coding--IMHO--to get EBS as a song option for bards using the new bard song schema without having to split HLA
tables and with minimal changes to existing assets. Once it was sorted, it was easy to tie in the RR changes and address the
compat issues more completely.