Resolving conflict with CtB (first pass)
As reported by RateLimited on Discord, IWDification fails when trying to make cosmetic patches to spells since the resources
are no longer from IWDification but from Check the Bodies. This is just a first pass to allow installation; a deeper dive
should follow to address warnings and/or other secondary issues.