Potential crash issues with MS1
ORC_MELEE or ORC_RANGE entries are in animate.ids by default. If another mod removes them, then IWDSpells returns -1 when it
tries to look them up, resulting in the orcs used in MS1 getting -1 as their animation value. This crashes the game when MS1 is
used. Since the creatures being copied already use the correct animations (more or less) we instead leave the animation
unchanged.

https://www.gibberlings3.net/forums/topic/35817-casting-summon-monsters-crashes-the-game/