More thorough 'tieflings != extraplanar' treatment
I had done an ad hoc removal of this when working on the oBG2 revisions for the compartmentalized postproduction changes. This is
a more thorough treatment to also remove this in the EEs as well, primarily via iwdfix.tph. Also corrected an oversight for oBG2
where Harm was not getting the same restrictions as the other Cause X Wounds spells.